% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicHTML.R
\name{basicHTML}
\alias{basicHTML}
\title{Convert to a Basic Custom styled HTML Document}
\usage{
basicHTML(
  css = system.file("rmarkdown/templates/basicHTML/resources/style.css", package =
    "rmdCustomHTML"),
  toc = TRUE,
  toc_Depth = 4,
  toc_float = TRUE,
  ...
)
}
\arguments{
\item{css}{One or more css files to include}

\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_float}{\code{TRUE} to float the table of contents to the left of the
main document content. Rather than \code{TRUE} you may also pass a list of
options that control the behavior of the floating table of contents. See the
\emph{Floating Table of Contents} section below for details.}

\item{...}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link[rmarkdown]{html_document_base}}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
with the custom basic theme.
}
\description{
Format for converting from R Markdown to Basic Custom styled HTML Document. Basic defaults are listed but can be updated in YAML.
}
\details{
See R Markdown documentation for details on yaml \link[rmarkdown]{metadata}.
}
\examples{
\dontrun{

library(rmarkdown)

# create a new draft presentation
rmarkdown::draft("report.Rmd", template = "basicHTML", package = "rmdCustomHTML")

# simple invocation
rmarkdown::render("report.Rmd", basicHTML())
}

}
